Move string ops str_ops.c
