+++ mypyc/lib-rt/str_ops.c
@@ -1,0 +1,60 @@
+// String primitive operations
+//
+// These are registered in mypyc.primitives.str_ops.
+
+#include <Python.h>
+#include "CPy.h"
+
+static PyObject *CPyStr_GetItem(PyObject *str, CPyTagged index) {
+    if (PyUnicode_READY(str) != -1) {
+        if (CPyTagged_CheckShort(index)) {
+            Py_ssize_t n = CPyTagged_ShortAsSsize_t(index);
+            Py_ssize_t size = PyUnicode_GET_LENGTH(str);
+            if ((n >= 0 && n >= size) || (n < 0 && n + size < 0)) {
+                PyErr_SetString(PyExc_IndexError, "string index out of range");
+                return NULL;
+            }
+            if (n < 0)
+                n += size;
+            enum PyUnicode_Kind kind = (enum PyUnicode_Kind)PyUnicode_KIND(str);
+            void *data = PyUnicode_DATA(str);
+            Py_UCS4 ch = PyUnicode_READ(kind, data, n);
+            PyObject *unicode = PyUnicode_New(1, ch);
+            if (unicode == NULL)
+                return NULL;
+
+            if (PyUnicode_KIND(unicode) == PyUnicode_1BYTE_KIND) {
+                PyUnicode_1BYTE_DATA(unicode)[0] = (Py_UCS1)ch;
+            }
+            else if (PyUnicode_KIND(unicode) == PyUnicode_2BYTE_KIND) {
+                PyUnicode_2BYTE_DATA(unicode)[0] = (Py_UCS2)ch;
+            } else {
+                assert(PyUnicode_KIND(unicode) == PyUnicode_4BYTE_KIND);
+                PyUnicode_4BYTE_DATA(unicode)[0] = ch;
+            }
+            return unicode;
+        } else {
+            PyErr_SetString(PyExc_IndexError, "string index out of range");
+            return NULL;
+        }
+    } else {
+        PyObject *index_obj = CPyTagged_AsObject(index);
+        return PyObject_GetItem(str, index_obj);
+    }
+}
+
+static PyObject *CPyStr_Split(PyObject *str, PyObject *sep, CPyTagged max_split)
+{
+    Py_ssize_t temp_max_split = CPyTagged_AsSsize_t(max_split);
+    if (temp_max_split == -1 && PyErr_Occurred()) {
+        PyErr_SetString(PyExc_OverflowError, "Python int too large to convert to C ssize_t");
+            return NULL;
+    }
+    return PyUnicode_Split(str, sep, temp_max_split);
+}
+
+/* This does a dodgy attempt to append in place  */
+static PyObject *CPyStr_Append(PyObject *o1, PyObject *o2) {
+    PyUnicode_Append(&o1, o2);
+    return o1;
+}
